[Web] Chore : :package: remove chatPage directory and moved its files to chatRoomPage directory #306
chatPage라고 잘못 명명된 폴더를 삭제하고 파일들 chatRoomPage로 이동
